refactor(CAS-168): split policies.py into a yosoi.policy package (per-stack modules)
Make policy its own module so pipeline stacks expand horizontally over time:

- yosoi/policy/_base.py  — shared trust lattice + validation primitives (no stack deps)
- yosoi/policy/crawl.py  — the crawl stack (Crawl* models, CrawlRuntimeConfig, presets)
- yosoi/policy/core.py   — Policy value object + resolve_crawl_policy/check_policy
- yosoi/policy/__init__.py — lazy PEP-562 re-exports

Dependency direction core -> crawl -> _base (no cycles); a future scrape stack
drops in beside crawl.py. Public surface unchanged: every 'from yosoi.policies
import X' rewritten to 'from yosoi.policy import X'; top-level ys.* exports intact.
