+++ yosoi/__init__.py
@@ -75,18 +75,18 @@ _LAZY: dict[str, str] = {
     # pipeline
     'Pipeline': 'yosoi.core.pipeline',
     # policies
-    'CrawlBudget': 'yosoi.policies',
-    'CrawlPolicy': 'yosoi.policies',
-    'CrawlRuntimeConfig': 'yosoi.policies',
-    'CrawlSafety': 'yosoi.policies',
-    'CrawlTarget': 'yosoi.policies',
-    'EscalationPolicy': 'yosoi.policies',
-    'Policy': 'yosoi.policies',
-    'PolicyCheck': 'yosoi.policies',
-    'SchedulerPolicy': 'yosoi.policies',
-    'check_policy': 'yosoi.policies',
-    'policy_arn': 'yosoi.policies',
-    'resolve_crawl_policy': 'yosoi.policies',
+    'CrawlBudget': 'yosoi.policy',
+    'CrawlPolicy': 'yosoi.policy',
+    'CrawlRuntimeConfig': 'yosoi.policy',
+    'CrawlSafety': 'yosoi.policy',
+    'CrawlTarget': 'yosoi.policy',
+    'EscalationPolicy': 'yosoi.policy',
+    'Policy': 'yosoi.policy',
+    'PolicyCheck': 'yosoi.policy',
+    'SchedulerPolicy': 'yosoi.policy',
+    'check_policy': 'yosoi.policy',
+    'policy_arn': 'yosoi.policy',
+    'resolve_crawl_policy': 'yosoi.policy',
     # generalization
     'PageFingerprint': 'yosoi.generalization.fingerprint',
     # integration transports
