+++ yosoi/__init__.pyi
@@ -28,17 +28,17 @@ from yosoi.models.snapshot import CacheVerdict as CacheVerdict
 from yosoi.models.snapshot import SelectorSnapshot as SelectorSnapshot
 from yosoi.models.snapshot import SnapshotMap as SnapshotMap
 from yosoi.models.snapshot import SnapshotStatus as SnapshotStatus
-from yosoi.policies import CrawlBudget as _CrawlBudget
-from yosoi.policies import CrawlPolicy as _CrawlPolicy
-from yosoi.policies import CrawlRuntimeConfig as _CrawlRuntimeConfig
-from yosoi.policies import CrawlSafety as _CrawlSafety
-from yosoi.policies import CrawlTarget as _CrawlTarget
-from yosoi.policies import EscalationPolicy as _EscalationPolicy
-from yosoi.policies import Outcome as _Outcome
-from yosoi.policies import Policy as _Policy
-from yosoi.policies import PolicyCheck as _PolicyCheck
-from yosoi.policies import SchedulerPolicy as _SchedulerPolicy
-from yosoi.policies import Trust as _Trust
+from yosoi.policy import CrawlBudget as _CrawlBudget
+from yosoi.policy import CrawlPolicy as _CrawlPolicy
+from yosoi.policy import CrawlRuntimeConfig as _CrawlRuntimeConfig
+from yosoi.policy import CrawlSafety as _CrawlSafety
+from yosoi.policy import CrawlTarget as _CrawlTarget
+from yosoi.policy import EscalationPolicy as _EscalationPolicy
+from yosoi.policy import Outcome as _Outcome
+from yosoi.policy import Policy as _Policy
+from yosoi.policy import PolicyCheck as _PolicyCheck
+from yosoi.policy import SchedulerPolicy as _SchedulerPolicy
+from yosoi.policy import Trust as _Trust
 from yosoi.types.field import Field as Field
 from yosoi.types.field import js as js
 from yosoi.types.registry import register_coercion as register_coercion
