+++ yosoi/core/pipeline/base.py
@@ -39,7 +39,7 @@ from yosoi.core.pipeline.utils import PipelineUtilsMixin
 from yosoi.core.verification import SelectorVerifier, SemanticValidator, field_rules_for_contract
 from yosoi.models.contract import Contract
 from yosoi.models.selectors import SelectorLevel
-from yosoi.policies import Policy
+from yosoi.policy import Policy
 from yosoi.storage import DebugManager, LLMTracker, SelectorStorage
 from yosoi.storage.discovery_strategy import DiscoveryStrategyStorage
 from yosoi.storage.js_scripts import JsScriptStorage
@@ -158,7 +158,7 @@ class Pipeline(
                 console for ``--json`` runs); a default themed console is built when omitted.
             keep_downloads: Keep downloaded files after the run (default). Set False to purge
                 the content-addressed blobs at run end while retaining provenance in index.json.
-            policy: Resolved pipeline :class:`~yosoi.policies.Policy` threaded from the API edge.
+            policy: Resolved pipeline :class:`~yosoi.policy.Policy` threaded from the API edge.
                 Stored once (``policy or Policy.from_env()``) as a forward-compat seam so future
                 policy-gated behavior reads ``self._policy`` instead of the environment; the
                 pipeline has no policy-gated branch today.
